# edge_api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DoroheDoro Edge API",
    "version": "0.2.0",
    "description": "Thin HTTP edge-api ingress for DoroheDoro. The service keeps HTTP routing, gRPC ingress, and NATS bridging in Go while domain responders stay behind request/reply subjects. This OpenAPI contract documents the HTTP surface exposed by the current MVP edge-api."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "listAgents",
        "summary": "List agents",
        "description": "Bridges an agent inventory list request to the `agents.list` NATS subject.",
        "responses": {
          "200": {
            "description": "Agent list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentsListResponse"
                },
                "example": {
                  "items": [
                    {
                      "id": "agent-01",
                      "host": "edge-host-01",
                      "status": "online",
                      "policy_id": "policy-linux-baseline",
                      "last_seen_at": "2026-03-20T12:00:00Z",
                      "labels": {
                        "env": "prod"
                      }
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/agents/{id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "getAgent",
        "summary": "Get agent by id",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                },
                "example": {
                  "item": {
                    "id": "agent-01",
                    "host": "edge-host-01",
                    "status": "online",
                    "policy_id": "policy-linux-baseline",
                    "last_seen_at": "2026-03-20T12:00:00Z",
                    "labels": {
                      "env": "prod"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "400": {
            "description": "Missing or invalid agent id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "invalid_argument",
                    "message": "agent id is required",
                    "request_id": "req-123"
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/agents/{id}/diagnostics": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "getAgentDiagnostics",
        "summary": "Get agent diagnostics",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Latest diagnostics snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDiagnosticsResponse"
                },
                "example": {
                  "item": {
                    "agent_id": "agent-01",
                    "status": "healthy",
                    "collected_at": "2026-03-20T12:00:00Z",
                    "checks": [
                      {
                        "name": "cpu",
                        "status": "ok",
                        "message": "within threshold"
                      }
                    ],
                    "payload": {
                      "cpu": 12,
                      "mem": 34
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "400": {
            "description": "Missing or invalid agent id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "invalid_argument",
                    "message": "agent id is required",
                    "request_id": "req-123"
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/policies": {
      "get": {
        "tags": [
          "policies"
        ],
        "operationId": "listPolicies",
        "summary": "List policies",
        "description": "Bridges a policy catalog list request to the `policies.list` NATS subject.",
        "responses": {
          "200": {
            "description": "Policy list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoliciesListResponse"
                },
                "example": {
                  "items": [
                    {
                      "id": "policy-linux-baseline",
                      "name": "Linux baseline",
                      "revision": "2026.03.20",
                      "description": "Baseline hardening policy for Linux nodes.",
                      "targets": [
                        "linux",
                        "amd64"
                      ],
                      "params": {
                        "enforce": "true"
                      }
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/policies/{id}": {
      "get": {
        "tags": [
          "policies"
        ],
        "operationId": "getPolicy",
        "summary": "Get policy by id",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Policy record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                },
                "example": {
                  "item": {
                    "id": "policy-linux-baseline",
                    "name": "Linux baseline",
                    "revision": "2026.03.20",
                    "description": "Baseline hardening policy for Linux nodes.",
                    "targets": [
                      "linux",
                      "amd64"
                    ],
                    "params": {
                      "enforce": "true"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "400": {
            "description": "Missing or invalid policy id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "invalid_argument",
                    "message": "policy id is required",
                    "request_id": "req-123"
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/deployments": {
      "get": {
        "tags": [
          "deployments"
        ],
        "operationId": "listDeployments",
        "summary": "List deployments",
        "responses": {
          "200": {
            "description": "Deployment jobs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentsListResponse"
                },
                "example": {
                  "items": [
                    {
                      "id": "dep-01",
                      "policy_id": "policy-linux-baseline",
                      "status": "accepted",
                      "created_at": "2026-03-20T12:00:00Z",
                      "agent_ids": [
                        "agent-01",
                        "agent-02"
                      ],
                      "params": {
                        "rollout": "canary"
                      }
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      },
      "post": {
        "tags": [
          "deployments"
        ],
        "operationId": "createDeployment",
        "summary": "Create deployment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentCreateRequest"
              },
              "example": {
                "policy_id": "policy-linux-baseline",
                "agent_ids": [
                  "agent-01",
                  "agent-02"
                ],
                "params": {
                  "rollout": "canary"
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Deployment request accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentAcceptedResponse"
                },
                "example": {
                  "accepted": true,
                  "item": {
                    "id": "dep-01",
                    "policy_id": "policy-linux-baseline",
                    "status": "accepted",
                    "created_at": "2026-03-20T12:00:00Z",
                    "agent_ids": [
                      "agent-01",
                      "agent-02"
                    ],
                    "params": {
                      "rollout": "canary"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "400": {
            "description": "Invalid deployment payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "invalid_argument",
                    "message": "policy_id is required",
                    "request_id": "req-123"
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/deployments/{id}": {
      "get": {
        "tags": [
          "deployments"
        ],
        "operationId": "getDeployment",
        "summary": "Get deployment by id",
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                },
                "example": {
                  "item": {
                    "id": "dep-01",
                    "policy_id": "policy-linux-baseline",
                    "status": "accepted",
                    "created_at": "2026-03-20T12:00:00Z",
                    "agent_ids": [
                      "agent-01",
                      "agent-02"
                    ],
                    "params": {
                      "rollout": "canary"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "400": {
            "description": "Missing or invalid deployment id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "invalid_argument",
                    "message": "deployment id is required",
                    "request_id": "req-123"
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/logs/search": {
      "post": {
        "tags": [
          "logs"
        ],
        "operationId": "searchLogs",
        "summary": "Search logs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogSearchRequest"
              },
              "example": {
                "query": "timeout OR error",
                "from": "2026-03-20T11:00:00Z",
                "to": "2026-03-20T12:00:00Z",
                "host": "edge-host-01",
                "service": "nginx",
                "severity": "warn",
                "limit": 100,
                "filters": {
                  "env": "prod"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogSearchResponse"
                },
                "example": {
                  "items": [
                    {
                      "timestamp": "2026-03-20T11:59:30Z",
                      "agent_id": "agent-01",
                      "host": "edge-host-01",
                      "service": "nginx",
                      "severity": "warn",
                      "message": "upstream timeout",
                      "labels": {
                        "env": "prod"
                      }
                    }
                  ],
                  "next_cursor": "logs:cursor:2",
                  "total": 1
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "400": {
            "description": "Invalid logs search payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "error": {
                    "code": "invalid_argument",
                    "message": "invalid JSON body",
                    "request_id": "req-123"
                  }
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/logs/histogram": {
      "get": {
        "tags": [
          "logs"
        ],
        "operationId": "getLogsHistogram",
        "summary": "Get logs histogram",
        "parameters": [
          {
            "$ref": "#/components/parameters/From"
          },
          {
            "$ref": "#/components/parameters/To"
          },
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Service"
          },
          {
            "$ref": "#/components/parameters/Severity"
          },
          {
            "$ref": "#/components/parameters/AgentId"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Query"
          },
          {
            "$ref": "#/components/parameters/Interval"
          }
        ],
        "responses": {
          "200": {
            "description": "Histogram buckets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistogramResponse"
                },
                "example": {
                  "items": [
                    {
                      "ts": "2026-03-20T11:55:00Z",
                      "count": 17
                    },
                    {
                      "ts": "2026-03-20T12:00:00Z",
                      "count": 25
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
    "/api/v1/logs/severity": {
      "get": {
        "tags": [
          "logs"
        ],
        "operationId": "getLogsSeverity",
        "summary": "Get severity aggregation",
        "parameters": [
          {
            "$ref": "#/components/parameters/From"
          },
          {
            "$ref": "#/components/parameters/To"
          },
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Service"
          },
          {
            "$ref": "#/components/parameters/Severity"
          },
          {
            "$ref": "#/components/parameters/AgentId"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Query"
          },
          {
            "$ref": "#/components/parameters/TopN"
          }
        ],
        "responses": {
          "200": {
            "description": "Severity aggregation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeverityResponse"
                },
                "example": {
                  "items": [
                    {
                      "name": "error",
                      "count": 3
                    },
                    {
                      "name": "warn",
                      "count": 12
                    }
                  ]
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "X-NATS-Subject": {
                "$ref": "#/components/headers/NATSSubject"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        }
      }
    },
...
    "/api/v1/stream/logs": {
      "get": {
        "tags": [
          "stream"
        ],
        "operationId": "streamLogs",
        "summary": "Stream logs over SSE",
        "description": "Subscribes the client to server-sent events backed by the `ui.stream.logs` NATS subject.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Host"
          },
          {
            "$ref": "#/components/parameters/Service"
          },
          {
            "$ref": "#/components/parameters/Severity"
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream established.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "event: log\ndata: {\"type\":\"log\",\"subject\":\"ui.stream.logs\",\"payload\":{\"timestamp\":\"2026-03-20T11:59:30Z\",\"service\":\"nginx\",\"severity\":\"warn\",\"message\":\"upstream timeout\"}}\n\n"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Id": {
        "in": "path",
        "name": "id",
        "required": true,
        "description": "Resource identifier.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "headers": {
      "RequestId": {
        "description": "Correlation identifier attached by the edge-api.",
        "schema": {
          "type": "string"
        }
      },
      "NATSSubject": {
        "description": "NATS subject used for the bridged upstream request.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "AgentsListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Agent"
            }
          },
          "next_cursor": {
            "type": "string",
            "example": "agents:cursor:2"
          }
        }
      },
      "AgentResponse": {
        "type": "object",
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/Agent"
          }
        }
      },
...
    },
    "responses": {
...
      "Unavailable": {
        "description": "Upstream bridge is unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
      "GatewayTimeout": {
        "description": "Upstream request timed out.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        }
      },
...
    }
  }
}